requirements.txt
fastapi
uvicorn[standard]
httpx[http2]
python-dotenv
pydantic